# openapi.json
{
  "components": {
    "schemas": {
...
      "Envelope_TeamResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TeamResponse"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[TeamResponse]",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "description": "도메인 에러 코드. 분기는 message 가 아니라 이 값으로 한다",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "PageMeta": {
        "properties": {
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "다음 페이지 커서. null 이면 마지막 페이지다. 다음 요청의 `cursor` 에 넣는다",
            "title": "Next Cursor"
          }
        },
        "required": [
          "next_cursor"
        ],
        "title": "PageMeta",
        "type": "object"
      },
...
      "Page_TeamResponse_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/TeamResponse"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Page[TeamResponse]",
        "type": "object"
      },
...
      "TeamCreate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "TeamCreate",
        "type": "object"
      },
...
      "TeamResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "format": "uuid",
            "title": "Created By",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "member_count": {
            "title": "Member Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "team_id": {
            "format": "uuid",
            "title": "Team Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "team_id",
          "name",
          "description",
          "created_by",
          "created_at",
          "updated_at",
          "member_count"
        ],
        "title": "TeamResponse",
        "type": "object"
      },
      "TeamUpdate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "TeamUpdate",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "성공 응답은 `{data, meta}`, 실패는 `{error: {code, message, details}}` 다. 분기는 HTTP 상태가 아니라 `error.code` 로 한다 — 같은 상태 코드에 코드가 여럿이다.\n\n입력 검증 실패는 전부 **400 `INVALID_INPUT`** 이며 `details.fields` 에 위치가 온다. 인증은 `Authorization: Bearer <access_token>`. 목록은 커서 페이지네이션이다: `meta.next_cursor` 를 다음 요청의 `cursor` 로 넘기고, null 이면 끝이다.",
    "title": "Prolog (StoryForge) API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/teams": {
      "get": {
        "description": "최근에 만들어진 팀부터. 커서 페이지네이션이다.",
        "operationId": "list_teams_v1_teams_get",
        "parameters": [
          {
            "description": "페이지 크기. 기본 20 · 최대 100",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "페이지 크기. 기본 20 · 최대 100",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "이전 응답의 `meta.next_cursor`. 첫 페이지는 비운다",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "이전 응답의 `meta.next_cursor`. 첫 페이지는 비운다",
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TeamResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_INPUT": {
                    "summary": "요청 값이 올바르지 않습니다",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "details": {},
                        "message": "요청 값이 올바르지 않습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVALID_INPUT` — 요청 값이 올바르지 않습니다"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "인증이 필요합니다",
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "details": {},
                        "message": "인증이 필요합니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `UNAUTHORIZED` — 인증이 필요합니다"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "내가 속한 팀 목록",
        "tags": [
          "TEAM"
        ]
      },
      "post": {
        "description": "만든 사람이 팀의 owner 가 된다.",
        "operationId": "create_team_v1_teams_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_TeamResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_INPUT": {
                    "summary": "요청 값이 올바르지 않습니다",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "details": {},
                        "message": "요청 값이 올바르지 않습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVALID_INPUT` — 요청 값이 올바르지 않습니다"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "인증이 필요합니다",
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "details": {},
                        "message": "인증이 필요합니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `UNAUTHORIZED` — 인증이 필요합니다"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "팀을 만든다",
        "tags": [
          "TEAM"
        ]
      }
    },
    "/v1/teams/{teamId}": {
      "delete": {
        "description": "owner 만. 팀 프로젝트가 하나라도 남아 있으면 `TEAM_HAS_ACTIVE_PROJECTS` 로 거부된다.",
        "operationId": "delete_team_v1_teams__teamId__delete",
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Teamid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_INPUT": {
                    "summary": "요청 값이 올바르지 않습니다",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "details": {},
                        "message": "요청 값이 올바르지 않습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVALID_INPUT` — 요청 값이 올바르지 않습니다"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "인증이 필요합니다",
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "details": {},
                        "message": "인증이 필요합니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `UNAUTHORIZED` — 인증이 필요합니다"
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "FORBIDDEN": {
                    "summary": "권한이 없습니다",
                    "value": {
                      "error": {
                        "code": "FORBIDDEN",
                        "details": {},
                        "message": "권한이 없습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `FORBIDDEN` — 권한이 없습니다"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "TEAM_NOT_FOUND": {
                    "summary": "팀을 찾을 수 없습니다",
                    "value": {
                      "error": {
                        "code": "TEAM_NOT_FOUND",
                        "details": {},
                        "message": "팀을 찾을 수 없습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `TEAM_NOT_FOUND` — 팀을 찾을 수 없습니다"
          },
          "409": {
            "content": {
              "application/json": {
                "examples": {
                  "TEAM_HAS_ACTIVE_PROJECTS": {
                    "summary": "팀에 프로젝트가 남아 있습니다",
                    "value": {
                      "error": {
                        "code": "TEAM_HAS_ACTIVE_PROJECTS",
                        "details": {},
                        "message": "팀에 프로젝트가 남아 있습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `TEAM_HAS_ACTIVE_PROJECTS` — 팀에 프로젝트가 남아 있습니다"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "팀을 삭제한다",
        "tags": [
          "TEAM"
        ]
      },
      "get": {
        "description": "팀원만 볼 수 있다. `member_count` 는 조회 시점의 인원이다.",
        "operationId": "get_team_v1_teams__teamId__get",
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Teamid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_TeamResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_INPUT": {
                    "summary": "요청 값이 올바르지 않습니다",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "details": {},
                        "message": "요청 값이 올바르지 않습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVALID_INPUT` — 요청 값이 올바르지 않습니다"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "인증이 필요합니다",
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "details": {},
                        "message": "인증이 필요합니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `UNAUTHORIZED` — 인증이 필요합니다"
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "FORBIDDEN": {
                    "summary": "권한이 없습니다",
                    "value": {
                      "error": {
                        "code": "FORBIDDEN",
                        "details": {},
                        "message": "권한이 없습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `FORBIDDEN` — 권한이 없습니다"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "TEAM_NOT_FOUND": {
                    "summary": "팀을 찾을 수 없습니다",
                    "value": {
                      "error": {
                        "code": "TEAM_NOT_FOUND",
                        "details": {},
                        "message": "팀을 찾을 수 없습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `TEAM_NOT_FOUND` — 팀을 찾을 수 없습니다"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "팀을 조회한다",
        "tags": [
          "TEAM"
        ]
      },
      "patch": {
        "description": "owner · admin 만. 보낸 필드만 바뀐다.",
        "operationId": "update_team_v1_teams__teamId__patch",
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Teamid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_TeamResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_INPUT": {
                    "summary": "요청 값이 올바르지 않습니다",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "details": {},
                        "message": "요청 값이 올바르지 않습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVALID_INPUT` — 요청 값이 올바르지 않습니다"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "인증이 필요합니다",
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "details": {},
                        "message": "인증이 필요합니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `UNAUTHORIZED` — 인증이 필요합니다"
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "FORBIDDEN": {
                    "summary": "권한이 없습니다",
                    "value": {
                      "error": {
                        "code": "FORBIDDEN",
                        "details": {},
                        "message": "권한이 없습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `FORBIDDEN` — 권한이 없습니다"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "TEAM_NOT_FOUND": {
                    "summary": "팀을 찾을 수 없습니다",
                    "value": {
                      "error": {
                        "code": "TEAM_NOT_FOUND",
                        "details": {},
                        "message": "팀을 찾을 수 없습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `TEAM_NOT_FOUND` — 팀을 찾을 수 없습니다"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "팀 정보를 수정한다",
        "tags": [
          "TEAM"
        ]
      }
    },
...
    "/v1/teams/{teamId}/invitations/{invitationId}": {
      "delete": {
        "description": "owner · admin 만. 대기 중(pending)인 초대만 취소할 수 있다.\n\n취소하면 같은 주소로 다시 초대할 수 있다.",
        "operationId": "revoke_invitation_v1_teams__teamId__invitations__invitationId__delete",
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Teamid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "invitationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invitationid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_INPUT": {
                    "summary": "요청 값이 올바르지 않습니다",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "details": {},
                        "message": "요청 값이 올바르지 않습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVALID_INPUT` — 요청 값이 올바르지 않습니다"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "인증이 필요합니다",
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "details": {},
                        "message": "인증이 필요합니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `UNAUTHORIZED` — 인증이 필요합니다"
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "FORBIDDEN": {
                    "summary": "권한이 없습니다",
                    "value": {
                      "error": {
                        "code": "FORBIDDEN",
                        "details": {},
                        "message": "권한이 없습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `FORBIDDEN` — 권한이 없습니다"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "TEAM_INVITATION_NOT_FOUND": {
                    "summary": "초대를 찾을 수 없습니다",
                    "value": {
                      "error": {
                        "code": "TEAM_INVITATION_NOT_FOUND",
                        "details": {},
                        "message": "초대를 찾을 수 없습니다"
                      }
                    }
                  },
                  "TEAM_NOT_FOUND": {
                    "summary": "팀을 찾을 수 없습니다",
                    "value": {
                      "error": {
                        "code": "TEAM_NOT_FOUND",
                        "details": {},
                        "message": "팀을 찾을 수 없습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `TEAM_INVITATION_NOT_FOUND` — 초대를 찾을 수 없습니다\n- `TEAM_NOT_FOUND` — 팀을 찾을 수 없습니다"
          },
          "409": {
            "content": {
              "application/json": {
                "examples": {
                  "INVITATION_NOT_PENDING": {
                    "summary": "이미 처리된 초대입니다",
                    "value": {
                      "error": {
                        "code": "INVITATION_NOT_PENDING",
                        "details": {},
                        "message": "이미 처리된 초대입니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVITATION_NOT_PENDING` — 이미 처리된 초대입니다"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "팀 초대를 취소한다",
        "tags": [
          "TEAM"
        ]
      }
    },
...
  }
}